accelerometer/Accelerometer.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Accelerometer Test</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://www.w3.org/TR/accelerometer/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/generic-sensor-tests.js"></script>
<div id="log"></div>
<iframe src="support-iframe.html" id="frame" style="display:none" sandbox="allow-scripts">
</iframe>
<script>

function verify_sensor_reading(reading) {
  return reading.x != 0 ||
         reading.y != 0 ||
         reading.z != 0;
}

async_test(t => {
  let sensor1 = new Accelerometer({includeGravity: false});
  let sensor2 = new Accelerometer();
  sensor1.onactivate = t.step_func_done(() => {
    assert_not_equals(sensor1.reading, sensor2.reading);
    sensor1.stop();
    sensor2.stop();
  });
  sensor1.onerror = t.step_func_done(event => {
    assert_unreached(event.error.name + ":" + event.error.message);
  });
  sensor2.onerror = t.step_func_done(event => {
    assert_unreached(event.error.name + ":" + event.error.message);
  });
  sensor1.start();
  sensor2.start();
}, "when construct Accelerometer with includeGravity that is set to false, the reading is different from the default one");

runSensorFrequency(Accelerometer);
runGenericSensorTests(Accelerometer, AccelerometerReading, verify_sensor_reading);

</script>
